move files to proper directories
