+++ test/http/response/Makefile
@@ -1,0 +1,23 @@
+NAME = response_test
+SRCS = main.cpp ../../../src/http/response/response.cpp
+OBJS = $(SRCS:.cpp=.o)
+
+# compiler
+CXX = c++
+CXXFLAGS = -Wall -Wextra -Werror -std=c++98
+
+# rules
+all: $(NAME)
+
+$(NAME): $(OBJS)
+	$(CXX) $(CXXFLAGS) -o $(NAME) $(OBJS)
+
+clean:
+	$(RM) $(OBJS)
+
+fclean: clean
+	$(RM) $(NAME)
+
+re: fclean all
+
+.PHONY: all clean fclean re
